% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtablefunction.R
\name{exportintotableraw}
\alias{exportintotableraw}
\title{This thing extracts all the relevant output from lm, such as the coefficients, standard error, confidence interval, and partial eta square}
\usage{
exportintotableraw(x)
}
\description{
This thing extracts all the relevant output from lm, such as the coefficients, standard error, confidence interval, and partial eta square
}
