% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtablefunction.R
\name{regtable}
\alias{regtable}
\title{This is the final function that produces the csv output. x is the lm object, and y is the name of the csv table.}
\usage{
regtable(x, y)
}
\description{
This is the final function that produces the csv output. x is the lm object, and y is the name of the csv table.
}
